fixed issue #6
After the reservation has been sent out to the property owner by guest, availability calendar for this unit has to be updated. Dates for which this reservation is for have to be changed in color, from green to light pink. These dates should not be available for booking. If owner accepts booking they should turn to red and if owner declines they should be turn back to green ans be available for booking.